Changing floats to doubles.

+++ TrafficDataContracts/src/main/java/fi/ahto/example/traffic/data/contracts/internal/ShapeData.java
@@ -27,7 +27,7 @@ public class ShapeData implements Serializable {
     @JsonProperty("Sequence")
     public int seq;
     @JsonProperty("Latitude")
-    public float latitude;
+    public double latitude;
     @JsonProperty("Longitude")
-    public float longitude;
+    public double longitude;
 }
